글 삭제에 인가 처리

+++ backend/src/main/java/com/back/p67260811/domain/post/post/controller/ApiV1PostController.java
@@ -128,8 +128,18 @@ public class ApiV1PostController {
 
     @DeleteMapping("/{id}")
     public RsData<Void> delete(
-            @PathVariable int id
+            @PathVariable int id,
+            @RequestHeader("Authorization") String apiKey
     ) {
+        String authorization = apiKey.replace("Bearer ", "");
+
+        Member actor = memberService.findByApiKey(authorization).orElseThrow(() -> new ServiceException("401-1", "API 키가 올바르지 않습니다."));
+
+        Post post = postService.findById(id).get();
+
+
+        if(!actor.equals(post.getAuthor())) throw new ServiceException("403-1", "삭제 권한이 없습니다.");
+
         postService.delete(id);
 
         return new RsData<>(
